requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/test.txt requirements/base.in requirements/test.in
#
attrs==20.3.0
    # via pytest
coverage==5.3.1
    # via pytest-cov
ddt==1.3.1
    # via -r requirements/test.in
iniconfig==1.1.1
    # via pytest
packaging==20.8
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.0
    # via -r requirements/test.in
pytest-django==4.1.0
    # via -r requirements/test.in
pytest==6.2.1
    # via
    #   pytest-cov
    #   pytest-django
toml==0.10.2
    # via pytest
